make LazyModule picklable
