+++ src/transformers/file_utils.py
@@ -1864,6 +1864,8 @@ class _BaseLazyModule(ModuleType):
     # https://github.com/optuna/optuna/blob/master/optuna/integration/__init__.py
     def __init__(self, name, import_structure):
         super().__init__(name)
+        self._name = name
+        self._import_structure = import_structure
         self._modules = set(import_structure.keys())
         self._class_to_module = {}
         for key, values in import_structure.items():
@@ -1890,6 +1892,28 @@ class _BaseLazyModule(ModuleType):
 
     def _get_module(self, module_name: str) -> ModuleType:
         raise NotImplementedError
+
+
+# This class should be in a transformers submodule in order to be properly picklable
+class _LazyModule(_BaseLazyModule):
+    """
+    Module class that surfaces all objects but only performs associated imports when the objects are requested.
+    """
+
+    __file__ = globals()["__file__"]
+    __path__ = [os.path.dirname(__file__)]
+
+    def _get_module(self, module_name: str):
+        return importlib.import_module("." + module_name, self.__name__)
+
+    def __getattr__(self, name: str):
+        # Special handling for the version, which is a constant from this module and not imported in a submodule.
+        if name == "__version__":
+            return __version__
+        return super().__getattr__(name)
+
+    def __reduce__(self):
+        return (self.__class__, (self._name, self._import_structure))
 
 
 def copy_func(f):
